Método save carrinho criado

+++ src/main/java/com/example/demo/entity/ItemCarrinho.java
@@ -21,6 +21,6 @@ public class ItemCarrinho {
     @ManyToOne
     @JoinColumn(name="carrinho_id", nullable=false)
     private Carrinho carrinho;
-    private Long quantidade;
+    private Integer quantidade;
     private BigDecimal preco;
 }
